requirements.txt
yfinance>=0.2.36
sqlalchemy>=2.0
streamlit>=1.32
plotly>=5.19
apscheduler>=3.10
pyyaml>=6.0
python-telegram-bot>=20.0
pandas>=2.0
